refactor: modularize command logic
