+++ project/templates/README.md
@@ -1,0 +1,1 @@
+### to be used for all standard html files for a standard file using a basic typescript/tailwind build
